requirements.txt
openai>=1.85.0
pydantic>=2.11.0
python-dotenv>=1.0.0
